cmd/client: split commands in multiple files
