+++ .tool/lint
@@ -14,7 +14,7 @@ for d in $(find . -type d -not -iwholename '*.git*' -a -not -iname '.tool' -a -n
 		--exclude='error return value not checked.*(Close|Log|Print).*\(errcheck\)$' \
 		--exclude='.*_test\.go:.*error return value not checked.*\(errcheck\)$' \
 		--exclude='duplicate of.*_test.go.*\(dupl\)$' \
-		--exclude='duplicate of.*main.go.*\(dupl\)$' \
+		--exclude='cmd\/client\/.*\.go.*\(dupl\)$' \
 		--disable=aligncheck \
 		--disable=gotype \
 		--disable=gas \
